refactor: finish LoaderManager class
